feat(cli): choses a new port if dev server is conflicted for apps
