+++ packages/sanity/src/_internal/cli/actions/app/devAction.ts
@@ -46,16 +46,20 @@ export default async function startAppDevServer(
   })
 
   try {
-    const spinner = output.spinner('Starting dev server').start()
-    await startDevServer({...config, spinner, skipStartLog: true, isApp: true})
+    output.print('Starting dev server')
+
+    const {server} = await startDevServer({...config, isApp: true})
+
+    const {port} = server.config.server
+    const httpHost = config.httpHost || 'localhost'
+
     const coreAppUrl = await getCoreAppURL({
       organizationId,
-      httpHost: config.httpHost,
-      httpPort: config.httpPort,
+      httpHost,
+      httpPort: port,
     })
-    spinner.succeed()
 
-    output.print(`Dev server started on port ${config.httpPort}`)
+    output.print(`Dev server started on port ${port}`)
     output.print(`View your app in the Sanity dashboard here:`)
     output.print(chalk.blue(chalk.underline(coreAppUrl)))
   } catch (err) {
