+++ packages/sanity/src/_internal/cli/server/devServer.ts
@@ -1,5 +1,5 @@
-import {type CliOutputter, type ReactCompilerConfig, type UserViteConfig} from '@sanity/cli'
-import chalk from 'chalk'
+import {type ReactCompilerConfig, type UserViteConfig} from '@sanity/cli'
+import {type ViteDevServer} from 'vite'
 
 import {debug} from './debug'
 import {extendViteConfigWithUserConfig, getViteConfig} from './getViteConfig'
@@ -14,23 +14,21 @@ export interface DevServerOptions {
   httpHost?: string
   projectName?: string
 
-  spinner: ReturnType<CliOutputter['spinner']>
   reactStrictMode: boolean
   reactCompiler: ReactCompilerConfig | undefined
   vite?: UserViteConfig
   entry?: string
   isApp?: boolean
-  skipStartLog?: boolean
 }
 
 export interface DevServer {
+  server: ViteDevServer
   close(): Promise<void>
 }
 
 export async function startDevServer(options: DevServerOptions): Promise<DevServer> {
   const {
     cwd,
-    spinner,
     httpPort,
     httpHost,
     basePath,
@@ -39,10 +37,8 @@ export async function startDevServer(options: DevServerOptions): Promise<DevServ
     reactCompiler,
     entry,
     isApp,
-    skipStartLog,
   } = options
 
-  const startTime = Date.now()
   debug('Writing Sanity runtime files')
   await writeSanityRuntime({cwd, reactStrictMode, watch: true, basePath, entry, isApp})
 
@@ -70,24 +66,12 @@ export async function startDevServer(options: DevServerOptions): Promise<DevServ
   debug('Creating vite server')
   const {createServer} = await import('vite')
   const server = await createServer(viteConfig)
-  const info = server.config.logger.info
 
   debug('Listening on specified port')
   await server.listen()
 
-  if (!skipStartLog) {
-    const startupDuration = Date.now() - startTime
-    const url = `http://${httpHost || 'localhost'}:${httpPort || '3333'}${basePath}`
-    const appType = isApp ? 'Sanity application' : 'Sanity Studio'
-
-    // Close the spinner before printing the message
-    spinner.succeed()
-    info(
-      `${appType} ` +
-        `using ${chalk.cyan(`vite@${require('vite/package.json').version}`)} ` +
-        `ready in ${chalk.cyan(`${Math.ceil(startupDuration)}ms`)} ` +
-        `and running at ${chalk.cyan(url)}`,
-    )
+  return {
+    server,
+    close: () => server.close(),
   }
-  return {close: () => server.close()}
 }
