+++ packages/sanity/src/_internal/cli/server/getViteConfig.ts
@@ -110,7 +110,10 @@ export async function getViteConfig(options: ViteOptions): Promise<InlineConfig>
     server: {
       host: server?.host,
       port: server?.port || 3333,
-      strictPort: true,
+      // Only enable strict port for studio,
+      // since apps can run on any port
+      strictPort: isApp ? false : true,
+
       /**
        * Significantly speed up startup time,
        * and most importantly eliminates the `new dependencies optimized: foobar. optimized dependencies changed. reloading`
